Add optional manual password and token for auth
Two fields have been added to the values.yaml of the chromadb-chart. The 'password' field for basic authentication and the 'value' field for token.
When these fields are not null, their corresponding authentication methods will use the provided string instead of generating a random string as a credential. They are null by default.
This change allows users to manually set their credentials and improve the system's flexibility.